Port i2pdotnet to .NET 6
