+++ I2P.NET.SAM/Connection/I2PSamConnection.cs
@@ -1,34 +1,13 @@
-﻿// This file is part of i2pdotnet.
-// Copyright (c) 2016
-//  
-// i2pdotnet is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as
-// published by the Free Software Foundation, version 3.
-//  
-// i2pdotnet is distributed in the hope that it will be useful, but
-// WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//  
-// You should have received a copy of the GNU Lesser General Public License
-// along with i2pdotnet.  If not, see <http://www.gnu.org/licenses/>.
-
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
+﻿using I2P.NET.SAM.Exceptions;
 using System.Net;
 using System.Net.Sockets;
 using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 
-namespace I2PNet
+namespace I2P.NET.SAM.Connection
 {
     internal class I2PSamConnection : IDisposable
     {
         private readonly int samPort;
-
         private TcpClient client;
         private StreamReader reader;
         private BinaryWriter writer;
