+++ I2P.NET.SAM/Sessions/I2PSession.cs
@@ -1,34 +1,15 @@
-﻿// This file is part of i2pdotnet.
-// Copyright (c) 2016
-//  
-// i2pdotnet is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as
-// published by the Free Software Foundation, version 3.
-//  
-// i2pdotnet is distributed in the hope that it will be useful, but
-// WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//  
-// You should have received a copy of the GNU Lesser General Public License
-// along with i2pdotnet.  If not, see <http://www.gnu.org/licenses/>.
-
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
+﻿using I2P.NET.SAM.Connection;
+using I2P.NET.SAM.Events;
+using I2P.NET.SAM.Interfaces;
 using System.Net;
 using System.Net.Sockets;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace I2PNet
+namespace I2P.NET.SAM.Sessions
 {
-    public class I2PSession : II2PSession
+    public class I2PSession : IBaseSession
     {
         private readonly int sessionId;
         private readonly I2PSamConnection controlSocket;
-        
         private bool isListeningForIncomingConnections;
 
         public I2PSession(int samPort, int? listenPort = null)
@@ -40,11 +21,8 @@ namespace I2PNet
         }
 
         public int SamPort { get; }
-
         public int? ListenPort { get; }
-
         public string LocalDestination { get; private set; }
-
         public event AcceptConnectionDelegate IncomingConnection;
 
         public async Task InitializeAsync()
@@ -90,11 +68,9 @@ namespace I2PNet
                     while (isListeningForIncomingConnections)
                     {
                         var client = await listener.AcceptTcpClientAsync();
-
                         var reader = new StreamReader(client.GetStream());
                         var status = reader.ReadLine();
                         string remoteDestination = status.Split(' ').First();
-
                         IncomingConnection?.Invoke(this, new AcceptConnectionEventArgs(client, remoteDestination));
                     }
                 }
