+++ I2P.NET.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32228.430
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "I2P.NET.Demo", "I2P.NET\I2P.NET.Demo.csproj", "{F1D6DCD2-FFDF-48F3-A3DB-63DBE4C0C6D0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "I2P.NET.SAM", "I2P.NET.SAM\I2P.NET.SAM.csproj", "{2CE26BB9-688C-4F40-BFF6-8DCCFB25FEDC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{16631391-9E91-48B2-8FB3-7E1DB3402A9F}"
+	ProjectSection(SolutionItems) = preProject
+		Manifesto.txt = Manifesto.txt
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F1D6DCD2-FFDF-48F3-A3DB-63DBE4C0C6D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1D6DCD2-FFDF-48F3-A3DB-63DBE4C0C6D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1D6DCD2-FFDF-48F3-A3DB-63DBE4C0C6D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1D6DCD2-FFDF-48F3-A3DB-63DBE4C0C6D0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2CE26BB9-688C-4F40-BFF6-8DCCFB25FEDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CE26BB9-688C-4F40-BFF6-8DCCFB25FEDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2CE26BB9-688C-4F40-BFF6-8DCCFB25FEDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2CE26BB9-688C-4F40-BFF6-8DCCFB25FEDC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {803051CA-C033-420D-9E2C-76100B4B5AE8}
+	EndGlobalSection
+EndGlobal
