--- i2pdotnet.sln
@@ -1,39 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26228.4
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{037BFA10-F5D5-47EE-B50E-B201ABBD0EA8}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{4E2A07CB-CF76-4E62-9174-F38EC4682B25}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "i2pdotnet", "src\i2pdotnet\i2pdotnet.csproj", "{F4297911-C03F-4B59-8EA7-43811F687E1C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleApp", "src\SampleApp\SampleApp.csproj", "{A62C6EBE-77FA-4CF2-BAF6-AA302E5048B1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F4297911-C03F-4B59-8EA7-43811F687E1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F4297911-C03F-4B59-8EA7-43811F687E1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F4297911-C03F-4B59-8EA7-43811F687E1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F4297911-C03F-4B59-8EA7-43811F687E1C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A62C6EBE-77FA-4CF2-BAF6-AA302E5048B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A62C6EBE-77FA-4CF2-BAF6-AA302E5048B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A62C6EBE-77FA-4CF2-BAF6-AA302E5048B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A62C6EBE-77FA-4CF2-BAF6-AA302E5048B1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{F4297911-C03F-4B59-8EA7-43811F687E1C} = {037BFA10-F5D5-47EE-B50E-B201ABBD0EA8}
-		{A62C6EBE-77FA-4CF2-BAF6-AA302E5048B1} = {037BFA10-F5D5-47EE-B50E-B201ABBD0EA8}
-	EndGlobalSection
-EndGlobal
